+++ kernel/ai-kernel/src/hmm.c
@@ -87,7 +87,7 @@ static int pool_grow(void){
 /* --------------------------- eviction ----------------------------- */
 static void page_writeback(hmm_page_t *p){
     if(!p->dirty) return;
-    kmemcpy((void*)(u32)p->host_addr,(void*)(u32)p->vram_addr,HMM_PAGE_SIZE);
+    kmemcpy((void*)(usize)p->host_addr,(void*)(usize)p->vram_addr,HMM_PAGE_SIZE);
     p->dirty=0; H.writebacks++;
 }
 static int evict_one(void){
@@ -128,7 +128,7 @@ static hmm_page_t *load_new(u32 m,u32 idx,int is_write){
     q->freq      = is_write?3:1;
     q->dirty     = is_write?1:0;
     q->resident  = 1;
-    kmemcpy((void*)(u32)va,(void*)(u32)q->host_addr,HMM_PAGE_SIZE);
+    kmemcpy((void*)(usize)va,(void*)(usize)q->host_addr,HMM_PAGE_SIZE);
     H.bytes_migrated += HMM_PAGE_SIZE;
     ht_insert(q);
     lru_push_mru(q);
